Adding attomics back into tutorial TSO uarch, also removing hard-coded path from runpipe.sh

+++ util/runpipe.sh
@@ -1,28 +1,28 @@
 #!/bin/bash
 
 # Exit on any errors
-# 1: path to tests
+# 2: path to tests
 # TODO: specify pipechek home
 
 set -e
-if [ -z "$1" ] || [ -z "$2" ] || [ -z "$3" ] || [ -z "$4" ] || [ -z "$5" ] || [ -z "$6" ]
+if [ -z "$1" ] || [ -z "$2" ] || [ -z "$3" ] || [ -z "$4" ] || [ -z "$5" ] || [ -z "$6" ] || [ -z "$7" ]
 then
-  echo "Usage: ./runpipe.sh <path to tests> <list of tests> <uarch name> <num threads> <test name> <path to uarch>"
-  exit 1
+  echo "Usage: ./runpipe.sh <path to pipecheck> <path to tests> <list of tests> <path to uarch> <uarch name> <num threads>"
+  exit 2
 fi
 
 # Store all of the tests in the given path into TESTS
-TESTS=$(cat $2)
-# OUTPUTDIR=results/graphs-herd-$4-$(date +"%m-%d-%y--%H-%M-%S-%p")/$6
-OUTPUTDIR=results/$4/$6
+TESTS=$(cat $3)
+# OUTPUTDIR=results/graphs-herd-$5-$(date +"%m-%d-%y--%H-%M-%S-%p")/$7
+OUTPUTDIR=results/$5/$7
 
 mkdir -p $OUTPUTDIR
 rm -f latest
 ln -s $OUTPUTDIR latest
-cp ~/mcm_verify/PipeCheck/src/pipecheck $OUTPUTDIR
+cp $1/pipecheck $OUTPUTDIR
 
 # Loop over each test
 date | tee $OUTPUTDIR/log.txt
-parallel -j$5 --results $OUTPUTDIR time $OUTPUTDIR/pipecheck -i $1/{} -o $OUTPUTDIR/{}.gv -m $3/$4 ::: $TESTS 2>&1 | grep "Stricter|BUG" # >  out.txt # || true
+parallel -j$6 --results $OUTPUTDIR time $OUTPUTDIR/pipecheck -i $2/{} -o $OUTPUTDIR/{}.gv -m $4/$5 ::: $TESTS 2>&1 | grep "Stricter|BUG" # >  out.txt # || true
 
 echo "$0 Done!"
